% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rosa.R
\name{barplot.rosa}
\alias{barplot.rosa}
\title{Barplot of block and component explained variances}
\usage{
\method{barplot}{rosa}(height, type = c("train", "CV"),
  ncomp = height$ncomp, col = mcolors(ncomp), ...)
}
\arguments{
\item{height}{A \code{rosa} object.}

\item{type}{Character indicating if explained variance should be based on training data ("train") or cross-validation ("CV").}

\item{ncomp}{Integer to control the number of components to plot (if fewer than the fitted number of components).}

\item{col}{Colors to use for the components.}

\item{...}{Additional parameters passed to \code{barplot}}
}
\value{
No return
}
\description{
Barplot of block and component explained variances
}
