% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.summaries.R
\name{blockexpl}
\alias{blockexpl}
\title{Explained variance (block-wise and component-wise)}
\usage{
blockexpl(object, ncomp = object$ncomp, type = c("train", "CV"))
}
\arguments{
\item{object}{A \code{rosa} object.}

\item{ncomp}{Integer to control the number of components to display (if fewer than the fitted number of components).}

\item{type}{Character indicating which type of explained variance to compute (default = "train", alternative = "CV").}
}
\value{
An object of class \code{rosaexpl} containing block-wise and component-wise explained variance contained in a matrix with attributes.
}
\description{
Explained variance (block-wise and component-wise)
}
