% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.methods.R
\name{coef.rosa}
\alias{coef.rosa}
\title{Regression coefficients from ROSA model.}
\usage{
\method{coef}{rosa}(object, ncomp = object$ncomp, comps, intercept = FALSE,
  ...)
}
\arguments{
\item{object}{A \code{rosa} object.}

\item{ncomp}{An \code{integer} vector giving the number of components to use for prediction.}

\item{comps}{An \code{integer} vector giving the exact components to use for predictions.}

\item{intercept}{A \code{logical} indicating if coefficients for the intercept should be included (default = FALSE).}

\item{...}{Additional arguments. Currently not implemented.}
}
\description{
Regression coefficients from ROSA model.
}
