% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_json}
\alias{demo_json}
\title{Dummy JSON for dummy dataset inputs}
\format{
An object of class \code{"list"}
}
\usage{
data(demo_json)
}
\description{
Dummy JSON for dummy dataset inputs
}
\examples{
data(demo_json)
demo_json
}
\keyword{datasets}
