% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{glossary}
\alias{glossary}
\title{Glossary and inputs documentation}
\format{
An object of class \code{"list"}
}
\usage{
data(glossary)
}
\description{
Glossary and inputs documentation
}
\examples{
data(glossary)
glossary
}
\keyword{datasets}
