% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rosa.R
\name{image.rosa}
\alias{image.rosa}
\title{Image plot for ROSA component interpretation.}
\usage{
\method{image}{rosa}(x, type = c("correlation", "residual", "order"),
  ncomp = x$ncomp, col = mcolors(128), legend = TRUE, mar = c(5, 6, 4,
  7), las = 1, zlim, main, ...)
}
\arguments{
\item{x}{A \code{rosa} object}

\item{type}{An optional \code{character} for selecting the plot type: "correlation" (default), "residual" or "order".}

\item{ncomp}{Integer to control the number of components to plot (if fewer than the fitted number of components).}

\item{col}{Colors used for the image plot, defaulting to mcolors(128).}

\item{legend}{Logical indicating if a legend should be included (default = TRUE).}

\item{mar}{Figure margins, default = c(5,6,4,7).}

\item{las}{Axis text direction, default = 1.}

\item{zlim}{A two element \code{numeric} limiting the color axis.}

\item{main}{The main title of the plot.}

\item{...}{Additional parameters passed to \code{image}}
}
\value{
No return.
}
\description{
Image plot for ROSA component interpretation.
}
