% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rosa.R
\name{loadingweightplot}
\alias{loadingweightplot}
\title{Loading weights plot}
\usage{
loadingweightplot(object, ylab = "loading weight", ...)
}
\arguments{
\item{object}{A \code{rosa} object.}

\item{ylab}{Character label to show on the y axis (default = "loading weight")}

\item{...}{Additional parameters passed to \code{loadingplot}}
}
\description{
Loading weights plot
}
