% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx.R
\name{mmm_pptx}
\alias{mmm_pptx}
\title{PowerPoint Presentation for Selected Model}
\usage{
mmm_pptx(
  InputCollect,
  OutputCollect,
  Extras = NULL,
  selected_model = NULL,
  template,
  master = "Rosa",
  filename = "RosaReport.pptx",
  ai_insights = FALSE,
  report_start = NULL,
  report_end = NULL,
  ...
)
}
\arguments{
\item{InputCollect, OutputCollect}{List. Robyn's output objects.}

\item{Extras}{List. Additional information required.
Import JSON elements from \code{robyn_write()} for quick build-up.}

\item{selected_model}{Character. In case there are more than one.}

\item{template, master}{Character. PPTX template file.}

\item{filename}{Character. Name of the output file.}

\item{ai_insights}{Boolean. Add AI Generated Insights?}

\item{report_start, report_end}{Character date. Reporting start and end dates.}

\item{...}{Additional parameters to pass to internal functions}
}
\description{
PowerPoint Presentation for Selected Model
}
