% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.methods.R
\name{predict.rosa}
\alias{predict.rosa}
\title{Predictions from ROSA models}
\usage{
\method{predict}{rosa}(object, newdata, ncomp = 1:object$ncomp, comps,
  type = c("response", "scores"), na.action = na.pass, ...)
}
\arguments{
\item{object}{A \code{rosa} object.}

\item{newdata}{Optional new data with the same types of predictor blocks as the ones used for fitting the object.}

\item{ncomp}{An \code{integer} vector giving the number of components to use for prediction.}

\item{comps}{An \code{integer} vector giving the exact components to use for predictions.}

\item{type}{A \code{character} indicating if responses or scores should be predicted (default = "response", or "scores")}

\item{na.action}{Function determining what to do with missing values in \code{newdata}.}

\item{...}{Additional arguments. Currently not implemented.}
}
\value{
Returns predicted responses or scores depending on inputs.
}
\description{
Predictions from ROSA models
}
