% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.summaries.R
\name{print.rosa}
\alias{print.rosa}
\title{Print method for ROSA}
\usage{
\method{print}{rosa}(x, ...)
}
\arguments{
\item{x}{An object fitted by \code{rosa}}

\item{...}{Unused.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for ROSA
}
