% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.summaries.R
\name{print.rosaexpl}
\alias{print.rosaexpl}
\title{Print function for explained variance}
\usage{
\method{print}{rosaexpl}(object, digits = 3, compwise = FALSE)
}
\arguments{
\item{object}{A \code{rosaexpl} object.}

\item{digits}{Integer number of digits to print.}

\item{compwise}{Logical indicating if block-wise (default/FALSE) or component-wise (TRUE) explained variance should be printed.}
}
\value{
No return
}
\description{
Print function for explained variance
}
