% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{rosa}
\alias{rosa}
\title{Analyse RNA-Seq counts data for spurious antisense}
\usage{
rosa(data, spikein_sense, spikein_anti, splice_sense, splice_anti, spike_ids,
  splice_ids, spike_lengths, splice_lengths, groups, resultdir,
  global = FALSE, xmin = 1e-12, xmax = 0.01, ymin = 1e-12,
  ymax = 1e-04, legendx = -12, legendy = 0)
}
\arguments{
\item{data}{A list of dataframes (1 for each replicate) containing sense and antisense counts by gene. 
Each dataframe should have columns: 
\itemize{
\item "Geneid" (the id of each gene)
\item "sense" (raw sense counts for each gene)
\item "anti" (raw antisense counts for each gene)
\item "tpmsense" (normalised sense counts (TPM))
\item "tpmanti" (normalised antisense counts (TPM))
\item "length" (length of gene)
}}

\item{spikein_sense}{A matrix of integers, where each column c corresponds to a replicate, 
and each row r corresponds to a spike in. Each entry is the sense counts for the r-th spike-in for replicate c.
The c-th column corresponds to the c-th entry in data. If there is no spike-in data, set to NULL.}

\item{spikein_anti}{A matrix of integers, where each column c corresponds to a replicate, 
and each row r corresponds to a spike in. Each entry is the antisense counts for the r-th spike-in for replicate c.
The c-th column corresponds to the c-th entry in data. If there is no spike-in data, set to NULL.}

\item{splice_sense}{A matrix of integers, where each column c corresponds to a replicate, 
and each row r corresponds to a gene. Each entry is the spliced sense counts (may be 0 or NA) for the r-th 
gene for replicate c. The c-th column corresponds to the c-th entry in data. If there is no spliced reads data, set to NULL.}

\item{splice_anti}{A matrix of integers, where each column c corresponds to a replicate, 
and each row r corresponds to a gene. Each entry is the spliced sense counts (may be 0 or NA) for the r-th 
gene for replicate c. The c-th column corresponds to the c-th entry in data. If there is no spliced reads data, set to NULL.}

\item{spike_ids}{A dataframe containing spike-in ids in column 'Geneid'}

\item{splice_ids}{A dataframe containing gene ids in column 'Geneid'}

\item{spike_lengths}{A dataframe containing spike-in lengths in column 'length'}

\item{splice_lengths}{A dataframe containing gene lengths in column 'length'}

\item{groups}{A character array listing each replicate by its position in data, and the spikein and 
spliced reads matrices. E.g. if replicates in data are in the order WT1,WT2,WT3,Mutant1,Mutant2,Mutant3 then 
groups should be set to c("WT1","WT2","WT3","Mutant1","Mutant2","Mutant3"). In future this parameter
will allow replicates to be combined in the analysis.}

\item{resultdir}{Full path to directory where plots should be output}

\item{global}{Use only spike-in ratios to calculate correction (Default \code{FALSE})}

\item{xmin}{Minimum x-value for plots of antisense vssense counts (Default \code{1e-12})}

\item{xmax}{Maximum x-value for plots of antisense vs sense counts (Default \code{1e-2})}

\item{ymin}{Minimum y-value for plots of antisense vs sense counts (Default \code{1e-12})}

\item{ymax}{Maximum y-value for plots of antisense vs sense counts (Default \code{1e-4})}

\item{legendx}{Location of left side of legend (Default \code{-12})}

\item{legendy}{Location of top of legend (Default \code{0})}
}
\value{
Returns a list with entries:
\itemize{
\item corrected_data (original data with corrected counts values),
\item spikeratios (ratios for the spike-in data), and 
\item spliceratios (ratios for the spliced reads data).
}
Each of spikeratios and spliceratios consists of a list with two entries: the first entry is a dataframe 
containing the global spike-in/spliced ratio of antisense:sense, and the standard deviation of the residuals 
for the associated linear model, for each condition listed in groups; the second entry is a list of dataframes,
one for each condition listed in groups, containing counts and ratios for each spike-in/gene.
}
\description{
RoSA calculates the incidence of spurious antisense counts in RNA-Seq data, using either (or both) antisense and
sense counts from spike-ins and spliced reads at known splice junctions.
}
\details{
This function was written with the intention of obtaining count and group parameters from an 
edgeR DGEList object d, via d$counts and d$samples$group. Input data from other sources may need to be adjusted
to match the required format.
}
