% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.summaries.R
\name{summary.rosa}
\alias{summary.rosa}
\title{Summary method for ROSA objects}
\usage{
\method{summary}{rosa}(object, what = c("all", "validation", "training"),
  digits = 4, print.gap = 2, ...)
}
\arguments{
\item{object}{A \code{rosa} object.}

\item{what}{A \code{character} indicating if summary should include all, validation or training.}

\item{digits}{The number of digits used for printing.}

\item{print.gap}{Gap between columns when printing.}

\item{...}{}
}
\value{
No return
}
\description{
Summary method for ROSA objects
}
