% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosa.R
\docType{package}
\name{Rosa}
\alias{Rosa-package}
\alias{Rosa}
\title{Interface for MMM powered by Robyn}
\description{
Powerful Marketing Mix Models (MMM) interface powered by Robyn. Seamlessly manage end-to-end projects, regardless of your coding expertise, enabling in-depth analysis, addressing crucial business queries on attribution, incrementality, and budget optimization. Rosa's dynamic interface harmonizes technical intricacies with user-friendly functionality, fostering collaboration and empowering users to navigate MMM complexities confidently. This innovative solution propels advertisers towards data-driven decision-making, providing invaluable insights into market dynamics and contributing to strategic decision-making worldwide.
}
\author{
\strong{Maintainer}: Bernardo Lares \email{laresbernardo@gmail.com}

}
